moved retina photo logic into proper app
